Refactored implementation of #959
