+++ PKHeX/Resources/text/en/LegalityCheckStrings_en.txt
